Сверстан экран "Детали вакансии", исправлено для detekt
