+++ app/src/main/java/ru/practicum/android/diploma/vacancy/domain/models/AddressEmployer.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.vacancy.domain.models
+
+data class AddressEmployer(
+    val city: String,
+    val street: String,
+    val building: String,
+    val raw: String
+)
