+++ app/src/main/java/ru/practicum/android/diploma/vacancy/domain/models/ContactsEmployer.kt
@@ -1,0 +1,7 @@
+package ru.practicum.android.diploma.vacancy.domain.models
+
+data class ContactsEmployer(
+    val name: String,
+    val email: String,
+    val phones: List<Phone>?
+)
