+++ app/src/main/java/ru/practicum/android/diploma/vacancy/domain/models/Employer.kt
@@ -1,0 +1,6 @@
+package ru.practicum.android.diploma.vacancy.domain.models
+
+data class Employer(
+    val name: String,
+    val logo: String
+)
