+++ app/src/main/java/ru/practicum/android/diploma/vacancy/domain/models/Phone.kt
@@ -1,0 +1,6 @@
+package ru.practicum.android.diploma.vacancy.domain.models
+
+data class Phone(
+    val comment: String?,
+    val formatted: String
+)
